+++ src/theme/components/image.ts
@@ -1,17 +1,32 @@
 interface CustomImageBaseStyle {
-  size: keyof typeof imageSizes;
+  ratio: keyof typeof imageRatios;
+  size: keyof typeof imageSizes | keyof typeof imageSizesBasedOnHeight;
+  sizeBasedOn: "height" | "width";
 }
 interface CustomImageWrapperBaseStyle {
   ratio: keyof typeof imageRatios;
-  size: keyof typeof imageSizes;
+  size: keyof typeof imageSizes | keyof typeof imageSizesBasedOnHeight;
+  sizeBasedOn: "height" | "width";
 }
 
-const imageWrap = {
-  position: "relative",
-  width: "100%",
-  height: "auto",
-  overflow: "hidden",
+const sizes = {
+  xxxsmall: 32,
+  xxsmall: 64,
+  xsmall: 96,
+  small: 165,
+  medium: 225,
+  large: 360,
 };
+
+const imageWrap = (sizeBasedOn: "height" | "width") => {
+  return {
+    position: "relative",
+    width: sizeBasedOn === "width" ? "100%" : "auto",
+    height: sizeBasedOn === "width" ? "auto" : "100%",
+    overflow: "hidden",
+  };
+};
+
 const sideMarginsAuto = {
   marginStart: "auto",
   marginEnd: "auto",
@@ -45,38 +60,85 @@ const imageSizes = {
     maxWidth: "360px",
   },
 };
+const imageSizesBasedOnHeight = {
+  default: {
+    maxHeight: "100%",
+  },
+  xxxsmall: {
+    ...sideMarginsAuto,
+    maxHeight: "32px",
+  },
+  xxsmall: {
+    ...sideMarginsAuto,
+    maxHeight: "64px",
+  },
+  xsmall: {
+    ...sideMarginsAuto,
+    maxHeight: "96px",
+  },
+  small: {
+    ...sideMarginsAuto,
+    maxHeight: "165px",
+  },
+  medium: {
+    ...sideMarginsAuto,
+    maxHeight: "225px",
+  },
+  large: {
+    ...sideMarginsAuto,
+    maxHeight: "360px",
+  },
+};
 const imageRatios = {
   fourByThree: {
-    ...imageWrap,
     paddingBottom: "75%",
   },
   oneByTwo: {
-    ...imageWrap,
     paddingBottom: "200%",
   },
-  original: {
-    ...imageWrap,
-  },
+  original: {},
   sixteenByNine: {
-    ...imageWrap,
     paddingBottom: "56.25%",
   },
   square: {
-    ...imageWrap,
     paddingBottom: "100%",
   },
   threeByFour: {
-    ...imageWrap,
     paddingBottom: "133.33%",
   },
   threeByTwo: {
-    ...imageWrap,
     paddingBottom: "66.66%",
   },
   twoByOne: {
-    ...imageWrap,
     paddingBottom: "50%",
   },
+};
+
+const imageWidthsBasedOnHeight = (height: number) => {
+  return {
+    fourByThree: {
+      maxWidth: `${(4 / 3) * height}px`,
+    },
+    oneByTwo: {
+      maxWidth: `${height / 2}px`,
+    },
+    original: {},
+    sixteenByNine: {
+      maxWidth: `${(16 / 9) * height}px`,
+    },
+    square: {
+      maxWidth: `${height}px`,
+    },
+    threeByFour: {
+      maxWidth: `${(3 / 4) * height}px`,
+    },
+    threeByTwo: {
+      maxWidth: `${(3 / 2) * height}px`,
+    },
+    twoByOne: {
+      maxWidth: `${2 * height}px`,
+    },
+  };
 };
 const CustomImage = {
   parts: ["captionWrappers"],
@@ -87,11 +149,20 @@ const CustomImage = {
       },
     },
   },
-  baseStyle: ({ size = "default" }: CustomImageBaseStyle) => ({
+  baseStyle: ({
+    ratio = "original",
+    size = "default",
+    sizeBasedOn = "width",
+  }: CustomImageBaseStyle) => ({
     figure: {
       margin: "auto",
-      width: "100%",
-      ...imageSizes[size],
+      ...(sizeBasedOn === "width"
+        ? { ...imageSizes[size], width: "100%" }
+        : {
+            ...imageSizesBasedOnHeight[size],
+            ...imageWidthsBasedOnHeight(sizes[size])[ratio],
+            height: "100%",
+          }),
       img: {
         backgroundColor: "ui.bg.default",
         marginBottom: "xxs",
@@ -109,8 +180,13 @@ const CustomImage = {
       boxSizing: "border-box",
       objectFit: "cover",
       position: "relative",
-      width: "100%",
-      ...imageSizes[size],
+      ...(sizeBasedOn === "width"
+        ? { ...imageSizes[size], width: "100%" }
+        : {
+            ...imageSizesBasedOnHeight[size],
+            ...imageWidthsBasedOnHeight(sizes[size])[ratio],
+            height: "100%",
+          }),
       _dark: {
         backgroundColor: "dark.ui.bg.default",
       },
@@ -122,22 +198,34 @@ const CustomImageWrapper = {
   baseStyle: ({
     ratio = "original",
     size = "default",
+    sizeBasedOn = "width",
   }: CustomImageWrapperBaseStyle) => ({
     marginStart: "auto",
     marginEnd: "auto",
-    width: "100%",
-    ...imageSizes[size],
+    ...(sizeBasedOn === "width"
+      ? { ...imageSizes[size], width: "100%" }
+      : {
+          ...imageSizesBasedOnHeight[size],
+          ...imageWidthsBasedOnHeight(sizes[size])[ratio],
+          height: "100%",
+        }),
     crop: {
+      ...imageWrap(sizeBasedOn),
       ...imageRatios[ratio],
+      ...(sizeBasedOn === "height" && {
+        ...imageWidthsBasedOnHeight(sizes[size])[ratio],
+      }),
     },
     img: {
+      ...(sizeBasedOn === "width"
+        ? { maxWidth: "100%" }
+        : { maxHeight: "100%" }),
       backgroundColor: "ui.bg.default",
       height: "100%",
       left: "0",
-      maxWidth: "100%",
+      width: "100%",
       position: "absolute",
       top: "0",
-      width: "100%",
       _dark: {
         backgroundColor: "dark.ui.bg.default",
       },